pure refactor, prepare for graph fixes
